chore: Added confirmation alert

+++ routes/user-account-routes.js
@@ -97,6 +97,13 @@ apiRouter.delete('/deleteBudget', (req, res) => {
     return;
   }
   const budgetId = req.body.budgetId;
+  if (!budgetId) {
+    console.error('No budget to delete');
+    return;
+  }
+  if (confirm('Are you sure you want to delete this budget?') === false) {
+    return;
+  }
   Budget.where({ _id: budgetId }).findOneAndDelete().then(() => {
     console.log('Budget Deleted');
     res.send('Budget Deleted');
